refactore calendar file structures
