+++ app/components/Header/Calendar/CalendarUI/CalendarUi.types.ts
@@ -1,0 +1,6 @@
+interface CalendarProps {
+  onDateSelect: (dates: {
+    checkIn: { day: number; month: object; year: number };
+    checkOut: { day: number; month: object; year: number };
+  } | null) => void;
+}
